myproject/requirements.txt
# Django 5.2.6
Django>=5.2,<6.0

# REST Framework
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg2-binary>=2.9,<3.0

# CORS
django-cors-headers>=4.3,<5.0

# Production server
gunicorn>=21.2,<22.0

# Security
django-environ>=1.0,<2.0